s/generateModelSId/generateLegacyModelSId
