+++ front/lib/utils.ts
@@ -8,24 +8,6 @@ export function classNames(...classes: string[]) {
   return classes.filter(Boolean).join(" ");
 }
 
-/**
- * Generates 10-character long model SId from [A-Za-z0-9] characters.
- */
-export function generateModelSId(prefix?: string): string {
-  const u = uuidv4();
-  const b = blake3(u, { length: 10 });
-  const sId = Buffer.from(b)
-    .map(uniformByteToCode62)
-    .map(alphanumFromCode62)
-    .toString();
-
-  if (prefix) {
-    return `${prefix}_${sId}`;
-  }
-
-  return sId;
-}
-
 export function client_side_new_id() {
   // blake3 is not available in the browser
   // remove the dashes from the uuid
@@ -277,41 +259,4 @@ export function sanitizeJSONOutput(obj: unknown): unknown {
     return sanitizedObj;
   }
   return obj;
-}
-
-/**
- * Given a code in between 0 and 61 included, returns the corresponding
- * character from [A-Za-z0-9]
- */
-function alphanumFromCode62(code: number) {
-  const CHAR_A = 65;
-  const CHAR_a = 97;
-  const CHAR_0 = 48;
-
-  if (code < 26) {
-    return CHAR_A + code;
-  }
-
-  if (code < 52) {
-    return CHAR_a + code - 26;
-  }
-
-  if (code < 62) {
-    return CHAR_0 + code - 52;
-  }
-
-  throw new Error("Invalid code");
-}
-
-/**
- * Given a byte, returns a code in between 0 and 61 included with a uniform
- * distribution guarantee, i.e. if the byte is uniformly drawn over 0-255, the
- * code will be uniformly drawn over 0-61.
- *
- * This is achieved by taking a modulo of 64 instead of 62, so the modulo is unbiased.
- * Then, if the result is 62 or 63, we draw a random number in [0, 61].
- */
-function uniformByteToCode62(byte: number): number {
-  const res = byte % 64;
-  return res < 62 ? res : Math.floor(Math.random() * 62);
 }
