+++ front/pages/api/poke/templates/index.ts
@@ -9,8 +9,8 @@ import type { NextApiRequest, NextApiResponse } from "next";
 
 import { USED_MODEL_CONFIGS } from "@app/components/providers/types";
 import { Authenticator, getSession } from "@app/lib/auth";
+import { generateLegacyModelSId } from "@app/lib/resources/string_ids";
 import { TemplateResource } from "@app/lib/resources/template_resource";
-import { generateModelSId } from "@app/lib/utils";
 import { apiError, withLogging } from "@app/logger/withlogging";
 import type { AssistantTemplateListType } from "@app/pages/api/w/[wId]/assistant/builder/templates";
 
@@ -103,7 +103,7 @@ async function handler(
         presetModelId: model.modelId,
         presetProviderId: model.providerId,
         presetTemperature: body.presetTemperature ?? null,
-        sId: generateModelSId(),
+        sId: generateLegacyModelSId(),
         tags: body.tags,
         visibility: body.visibility,
       });
