+++ front/pages/api/stripe/webhook.ts
@@ -22,8 +22,8 @@ import {
 import { maybeCancelInactiveTrials } from "@app/lib/plans/subscription";
 import { countActiveSeatsInWorkspace } from "@app/lib/plans/usage/seats";
 import { frontSequelize } from "@app/lib/resources/storage";
+import { generateLegacyModelSId } from "@app/lib/resources/string_ids";
 import { ServerSideTracking } from "@app/lib/tracking/server";
-import { generateModelSId } from "@app/lib/utils";
 import { renderLightWorkspaceType } from "@app/lib/workspace";
 import logger from "@app/logger/logger";
 import { apiError, withLogging } from "@app/logger/withlogging";
@@ -220,7 +220,7 @@ async function handler(
 
               await Subscription.create(
                 {
-                  sId: generateModelSId(),
+                  sId: generateLegacyModelSId(),
                   workspaceId: workspace.id,
                   planId: plan.id,
                   status: "active",
