+++ front/pages/api/w/[wId]/apps/index.ts
@@ -4,7 +4,7 @@ import type { NextApiRequest, NextApiResponse } from "next";
 
 import { Authenticator, getSession } from "@app/lib/auth";
 import { App } from "@app/lib/models/apps";
-import { generateModelSId } from "@app/lib/utils";
+import { generateLegacyModelSId } from "@app/lib/resources/string_ids";
 import logger from "@app/logger/logger";
 import { apiError, withLogging } from "@app/logger/withlogging";
 
@@ -78,7 +78,7 @@ async function handler(
       const description = req.body.description ? req.body.description : null;
 
       const app = await App.create({
-        sId: generateModelSId(),
+        sId: generateLegacyModelSId(),
         name: req.body.name,
         description,
         visibility: req.body.visibility,
